pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sparse-capability-transplants"
version = "0.1.0"
description = "Code and core docs for the Sparse Capability Transplants research project."
readme = "README.md"
requires-python = ">=3.11,<3.13"
dependencies = [
  "matplotlib>=3.8,<4",
  "PyYAML>=6.0,<7",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3,<9",
  "ruff>=0.11,<0.12",
]

[project.scripts]
sct = "src.cli:main"

[tool.setuptools.packages.find]
include = ["src*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
norecursedirs = ["tests/_tmp", "tests/__pycache__"]

[tool.ruff]
extend-exclude = ["artifacts", "runs", "tests/_tmp"]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B"]
